Start to remove first_image stuff
