+++ src/models/posts.rs
@@ -1,38 +1,6 @@
 use chrono::{DateTime, Utc};
-use serde::{Serialize, Deserialize};
+use serde::Deserialize;
 use sqlx::PgPool;
-
-
-pub struct Image {
-    pub filename: String,
-    pub thumbnail_filename: Option<String>,
-    pub width: i16,
-    pub height: i16,
-    pub kb: Option<i32>,
-}
-
-impl Image {
-    pub async fn create(pool: &PgPool, post_key: &str, props: Image) -> Result<(), String> {
-        let image_id = sqlx::query!(r#"
-            insert into image (filename, thumbnail_filename, width, height, kb)
-                values ($1, $2, $3, $4, $5)
-                returning id
-        "#, props.filename, props.thumbnail_filename, props.width, props.height, props.kb)
-            .fetch_one(pool)
-            .await
-            .map_err(|e| e.to_string())?;
-
-        sqlx::query!(r#"
-            insert into post_image (post_id, image_id)
-                values ((select id from post where key = $1), $2)
-        "#, post_key.to_owned(), image_id.id)
-            .execute(pool)
-            .await
-            .map(|_| ())
-            .map_err(|e| e.to_string())
-    }
-}
-
 
 #[derive(Deserialize)]
 pub struct PostProps {
@@ -44,25 +12,15 @@ pub struct PostPreview {
     pub author: Option<String>,
     pub key: String,
     pub title: String,
-    pub first_image: Option<String>,
     pub created_at: DateTime<Utc>,
+    pub thumbnail: Option<String>,
+    pub alt_text: Option<String>,
 }
 
 impl PostPreview {
     pub async fn load_latest(pool: &PgPool) -> Result<Vec<Self>, String> {
-        sqlx::query_as!(Self, r#"
-            select
-                author.name as author,
-                key,
-                title,
-                created_at,
-                first_image(content)
-            from post
-            left join author on author.id = post.author_id
-            where published
-            order by created_at desc
-            limit 4
-        "#).fetch_all(pool).await.map_err(|e| e.to_string())
+        sqlx::query_as!(Self, "select * from post_preview_vw limit 4")
+            .fetch_all(pool).await.map_err(|e| e.to_string())
     }
 }
 
@@ -97,7 +55,8 @@ pub struct Post {
     pub published: bool,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
-    pub first_image: Option<String>
+    pub thumbnail: Option<String>,
+    pub alt_text: Option<String>,
 }
 
 impl Post {
@@ -136,20 +95,8 @@ impl Post {
     }
 
     pub async fn load(pool: &PgPool, key: String) -> Result<Self, String> {
-        sqlx::query_as!(Self, "
-            select
-                author.name as author,
-                post.key,
-                post.title,
-                post.content,
-                post.published,
-                post.created_at,
-                post.updated_at,
-                first_image(post.content)
-
-            from post left join author on author.id = post.author_id
-            where key = $1
-        ", key).fetch_one(pool).await.map_err(|e| e.to_string())
+        sqlx::query_as!(Self, "select * from post_detail_vw where key = $1", key)
+            .fetch_one(pool).await.map_err(|e| e.to_string())
     }
 
     pub async fn delete(pool: &PgPool, key: &str) -> Result<(), String> {
@@ -159,39 +106,6 @@ impl Post {
             .map(|_| ())
             .map_err(|e| e.to_string())
     }
-}
-
-#[derive(Clone)]
-pub struct Author {
-    pub id: i32,
-    pub name: String,
-    pub email: String,
-    pub password_hash: String,
-}
-
-impl Author {
-    pub async fn load(pool: &PgPool, email: String) -> Result<Self, String> {
-        sqlx::query_as!(
-            Self,
-            "select id, email, name, password_hash from author where email = $1",
-            email,
-        ).fetch_one(pool).await.map_err(|e| e.to_string())
-    }
-}
-
-impl Into<AuthorWithoutPassword> for Author {
-    fn into(self) -> AuthorWithoutPassword {
-        let Author { id, name, email, .. } = self;
-
-        AuthorWithoutPassword { id, name, email }
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct AuthorWithoutPassword {
-    pub id: i32,
-    pub name: String,
-    pub email: String,
 }
 
 pub trait TitleSlug {
